encryted creditcard

+++ CreditCard.Application/CreditCard/Services/CreditCardService.cs
@@ -35,6 +35,9 @@ namespace CreditCard.Application.CreditCard.Services
 
         public async Task Delete(Guid id)
         {
+            if (id == null)
+                throw new Exception("No se pudo eliminar la tarjeta de crédito");
+
             await _creditCardRepository.DeleteAsync(id);
         }
 
@@ -58,7 +61,7 @@ namespace CreditCard.Application.CreditCard.Services
             cr.Id = creditCard.Id;
             cr.ClientId = creditCard.ClientId;
             cr.AccountNumber = creditCard.AccountNumber;
-            cr.CardNumber = creditCard.CardNumber;
+            cr.CardNumber = EncryptCreditCardNumber(creditCard.CardNumber);
             cr.ExpirationDate = creditCard.ExpirationDate;
             cr.CutoffDate = creditCard.CutoffDate;
             cr.PaymentDueDate = creditCard.PaymentDueDate;
@@ -69,13 +72,29 @@ namespace CreditCard.Application.CreditCard.Services
             cr.CashAdvance = creditCard.CashAdvance;
             cr.BalanceToDate = creditCard.BalanceToDate;
             cr.BalanceToCut = creditCard.BalanceToCut;
+            cr.Status = creditCard.Status;
 
             return cr;
+        }
+
+        public string EncryptCreditCardNumber(long creditCardNumber)
+        {
+            string creditCardString = creditCardNumber.ToString();
+            if (creditCardString.Length != 16)
+            {
+                throw new ArgumentException("El número de tarjeta de crédito debe tener 16 dígitos");
+            }
+
+            string encryptedCreditCard = new string('*', 12) + creditCardString.Substring(12);
+            return encryptedCreditCard;
         }
 
         public async Task Update(Guid id, UpdateCreditCardDto creditCardDto, CancellationToken cancellationToken)
         {
             CreditCards creditCards = await _creditCardRepository.GetByIdAsync(id);
+
+            if(!creditCards.Status)
+                throw new Exception("No se puede actualizar la tarjeta porque está desactivada");
 
             _mapper.Map(creditCardDto, creditCards);
 
@@ -107,6 +126,9 @@ namespace CreditCard.Application.CreditCard.Services
             if (creditCard == null)
                 throw new Exception("Tarjeta de crédito no encontrada");
 
+            if (!creditCard.Status)
+                throw new Exception("No se puede realizar el avance de efectivo porque la tarjeta está desactivada");
+
             // Calcula el cargo adicional y lo convierte a long
             var additionalCharge = (long)(creditCardCashAdvanceDto.CashAdvance * 0.05m);
 
@@ -120,13 +142,13 @@ namespace CreditCard.Application.CreditCard.Services
             creditCard.AvailableWithOverdraft -= creditCardCashAdvanceDto.CashAdvance + additionalCharge;
             creditCard.BalanceToDate = creditCard.CashAdvance;
 
-            await _creditCardRepository.TransferCashAdvance(creditCard);
+            await _creditCardRepository.TransferCashAdvanceAsync(creditCard);
         }
 
 
         public async Task<List<CreditCardResponseDto>> GetCreditCardByClientId(int clientId)
         {
-            var creditCards = await _creditCardRepository.GetCreditCardByClientId(clientId);
+            var creditCards = await _creditCardRepository.GetCreditCardByClientIdAsync(clientId);
 
             if (!creditCards.Any())
                 throw new Exception("No se pudo encontrar una tarjeta asociada a este cliente");
@@ -139,7 +161,7 @@ namespace CreditCard.Application.CreditCard.Services
                 cr.Id = creditCard.Id;
                 cr.ClientId = creditCard.ClientId;
                 cr.AccountNumber = creditCard.AccountNumber;
-                cr.CardNumber = creditCard.CardNumber;
+                cr.CardNumber = EncryptCreditCardNumber(creditCard.CardNumber);
                 cr.ExpirationDate = creditCard.ExpirationDate;
                 cr.CutoffDate = creditCard.CutoffDate;
                 cr.PaymentDueDate = creditCard.PaymentDueDate;
@@ -150,6 +172,7 @@ namespace CreditCard.Application.CreditCard.Services
                 cr.CashAdvance = creditCard.CashAdvance;
                 cr.BalanceToDate = creditCard.BalanceToDate;
                 cr.BalanceToCut = creditCard.BalanceToCut;
+                cr.Status = creditCard.Status;
 
                 creditCardsList.Add(cr);
             }
@@ -157,5 +180,16 @@ namespace CreditCard.Application.CreditCard.Services
             return creditCardsList;
         }
 
+        public async Task ChangeCreditCardStatus(Guid id, ChangeCreditCardStatusDto changeCreditCardStatusDto)
+        {
+            var creditCard = await _creditCardRepository.GetByIdAsync(id);
+
+            if (creditCard == null)
+                throw new Exception("Tarjeta de crédito no encontrada");
+
+            creditCard.Status = changeCreditCardStatusDto.Status;
+
+            await _creditCardRepository.ChangeCreditCardStatusAsync(creditCard);
+        }
     }
 }
